pyproject.toml
[project]
name = "blog-backend"
version = "0.1.0"
description = "Blog backend API with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.3.0",
    "bcrypt==4.0.1",
    "psutil>=7.1.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
